more exploring

+++ algorithm/src/cattells.py
@@ -3,4 +3,4 @@
 import numpy as np
 
 def read_16pf_as_ndarray():
-    return np.genfromtxt('../data/16pf.csv', dtype=int, skip_header=1, usecols=range(163))
+    return np.genfromtxt('../data/16pf.csv', dtype=float, skip_header=1, usecols=range(163))
